+++ CHANGES.rst
@@ -27,6 +27,8 @@ Changelog
   [sgeulette]
 - Added seal column on @@parapheo.
   [chris-adam]
+- Added external watchers for esign sessions.
+  [chris-adam, sgeulette]
 
 1.0a2 (2026-02-06)
 ------------------